refactor(cli): split evaluation pipeline into focused modules
- Keep orchestrator as a thin composition root for mode dispatch,

  aggregation, formatter selection, and final reporting

- Extract per-file evaluation, prompt-result processing, violation

  reporting, and agent-mode flow into dedicated modules

- Preserve runtime behavior while reducing cross-cutting coupling

  and making future changes easier to test and review
